requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.1

# Database
aiosqlite==0.19.0

# File Operations
aiofiles==23.2.1

# Data Validation
pydantic==2.5.0

# Task Queue & Caching
redis==5.0.1
celery==5.3.4
kombu==5.3.4

# WebSocket
websockets==12.0

# Logging
loguru==0.7.2

# Monitoring
prometheus-client==0.19.0

# Utils
python-dotenv==1.0.0
